fix(flywheel): split substrate-bug-weights route, add missing telemetry imports, update no-loss matrix
- Extract GET /api/flywheel/substrate-bug-weights into its own route file
  to keep flywheel.ts at its 1005-line baseline.
- Import parseFlywheelStatsWindow and computeFlywheelStats in flywheel.ts
  for the stats payload route.
- Add the new route to the G3 no-loss matrix.
