+++ tests/unit/lib/overdeck/no-loss-matrix.ts
@@ -264,6 +264,7 @@ export const NO_LOSS_MATRIX: MatrixEntry[] = [
   { surface: 'POST /api/flywheel/uat-generations/:name/promote',    kind: 'http', disposition: 'WRITE',       door: 'MergeWriter.promoteUat' },
   { surface: 'POST /api/flywheel/assemble-uat',                     kind: 'http', disposition: 'WRITE',       door: 'MergeWriter.assembleUat' },
   { surface: 'GET /api/flywheel/state',                             kind: 'http', disposition: 'AGGREGATE',   door: 'SettingsResolver.getFlywheelRuntime + run-state' },
+  { surface: 'GET /api/flywheel/substrate-bug-weights',             kind: 'http', disposition: 'AGGREGATE',   door: 'SubstrateBugWeightsService.listSubstrateBugWeights + FlywheelStats' },
 
   // ── hooks.ts ──────────────────────────────────────────────────────────────
   { surface: 'POST /api/memory/inject',                   kind: 'http', disposition: 'READ',        door: 'MemoryResolver.injectPromptTime' },
